feat : Feature/role based access control main (#736)
* feat: role based access control create a feature branch and add permissions objects and functions (#719)

* feat: added backend validation for features

* refactor: updated file name

* refactor: updated backend and frontend function

* refactor: permission object

* refactor: removed console

* refactor: updated return

* refactor: updated return

* fix: test cases of workspace and setting

* refactor: updated test cases

* refactor: test case

* fix: resolve comments

* 743 develop 403 forbidden page (#754)

* chore: added 403 unauthorized route

* refactor: updated import name

* style: updated styling of 403 and 404 page

* 711 apply access checks for results (#741)

* refactor: getAllCandidateTest function

* chore: added backend checks and loader function checks for results permission

* refactor: error response / created 404 page

* refactor: resolved comments

* refactor: updated constant name

* refactor: updated function naming

* refactor: resolved comments

* refactor: file name

* chore: added 403 unauthorized route

* refactor: function naming

* refactor: function naming

* refactor: updated tests

* refactor: test cases

* refactor: test cases

* refactor: updated test cases

* refactor: updated test cases

* refactor: updated test cases

* refactor: test cases

* 740 apply access checks for candidate invite (#745)

* refactor: getAllCandidateTest function

* chore: added backend checks and loader function checks for results permission

* chore: added backend access permission for invite candidate

* refactor: error response / created 404 page

* refactor: added unauthorized route in redirect

* refactor: added backend checks for candidate invite feature

* refactor: resolved comments

* refactor: updated constant name

* refactor: updated function naming

* refactor: resolved comments

* refactor: file name

* chore: added 403 unauthorized route

* refactor: function naming

* refactor: function naming

* refactor: resolved comments

* refactor: added workspace backend permissions (#775)

* refactor: added assessment permission to backend (#776)

* refactor: added tests permission checks to backend (#779)

* fix: resolve test cases for viewport issues

* refactor: added backend permission check for member features (#777)

* fix: test cases

* fix: test cases

* fix: test cases

* fix: test cases

* fix: test cases for smoke

* test: remove unnecessary code

* fix: test cases for result details

* fix: test cases of resultDetails

* fix: test cases

* fix: test cases

* fix: test cases

* fix: test cases of results details

* refactor: test case

* refactor: test cases

* refactor: test cases

* refactor: test case

* refactor: test case

* refactor: test cases

* 747 FE apply access checks for results (#781)

* refactor: added FE results permission check

* refactor: updated state

* refactor: added member FE checks (#783)

* refactor: added FE permission checks for workspace (#788)

* refactor: added FE checks for assessment (#790)

* 752 FE apply access checks for candidate invite (#782)

* refactor: added FE checks for candidate invite

* refactor: updated check

* refactor: test case

* refactor: updated checks

* 749 FE apply access checks for tests (#794)

* refactor: access checks for tests

* fix: resolve comments

* fix: resolve comments

---------


* chore: created test cases for logout

* refactor: test cases

* refactor: test case

* small fix

* refactor: added 3 dots check for recruiter (#808)

* Feat(): update members page according to new design (#814)

* refactor: updated tabs component

* refactor: updated tabs component

* refactor: updated test cases

* refactor: updated test case

* refactor: test cases

* refactor: updated test case

* refactor: updated prop name

---------

+++ app/models/user.server.ts
@@ -4,6 +4,7 @@ import { faker } from "@faker-js/faker"
 import type { Password, User } from "@prisma/client"
 import bcrypt from "bcryptjs"
 
+import { checkFeatureAuthorization } from "./authorization.server"
 import { sendMail, sendNewPassword } from "./sendgrid.servers"
 
 import { prisma } from "~/db.server"
@@ -17,31 +18,43 @@ export async function getUserById(id: User["id"]) {
 export async function deleteUserById(
   userId: string,
   workspaceId: string,
-  email: string
+  email: string,
+  id: string
 ) {
-  const deleteUserWorkspace = await prisma.userWorkspace.deleteMany({
-    where: { userId, workspaceId },
-  })
-  const invitedIdByEmail = await prisma.invites.findMany({
-    where: {
-      email: email,
-      workspaceId,
-      deleted: false,
-    },
-    select: {
-      id: true,
-    },
-  })
-  await prisma.invites.update({
-    where: {
-      id: invitedIdByEmail[0].id,
-    },
-    data: {
-      deleted: true,
-      deletedAt: new Date().toString(),
-    },
-  })
-  return deleteUserWorkspace
+  try {
+    if (
+      !(await checkFeatureAuthorization(id, workspaceId, "member", "delete"))
+    ) {
+      throw {
+        status: 403,
+      }
+    }
+    const deleteUserWorkspace = await prisma.userWorkspace.deleteMany({
+      where: { userId, workspaceId },
+    })
+    const invitedIdByEmail = await prisma.invites.findMany({
+      where: {
+        email: email,
+        workspaceId,
+        deleted: false,
+      },
+      select: {
+        id: true,
+      },
+    })
+    await prisma.invites.update({
+      where: {
+        id: invitedIdByEmail[0].id,
+      },
+      data: {
+        deleted: true,
+        deletedAt: new Date().toString(),
+      },
+    })
+    return deleteUserWorkspace
+  } catch (error) {
+    throw error
+  }
 }
 
 export async function getUserByEmail(email: User["email"]) {
